should improve circle menuitem and reduce some compiler ewarnings

+++ include/mymenu/menuitems_outputselectorcontrol.h
@@ -116,7 +116,7 @@ class OutputSelectorControl : public SelectorControl<int_least16_t> {
     virtual void setter(int_least16_t new_value) override {
         //if (this->debug) Serial.printf(F("ParameterSelectorControl changing from %i to %i\n"), this->actual_value_index, new_value);
         selected_value_index = actual_value_index = new_value;
-        if(new_value>=0 && new_value<this->available_objects->size() && this->target_object!=nullptr && this->setter_func!=nullptr) {
+        if(new_value>=0 && new_value<(int_least16_t)this->available_objects->size() && this->target_object!=nullptr && this->setter_func!=nullptr) {
             (this->target_object->*this->setter_func)(this->available_objects->get(new_value));
         }
     }
